requirements.txt
streamlit>=1.37
spotipy>=2.24
python-dotenv>=1.0
requests>=2.31
